Adding a rake task to let us patch up legacy_id values
For some reason the legacy_id of a bunch of People in 2023 events was
not populated. This breaks the video publishing system since it doesn't
know who to email the permission form to. As a fix, this task checks a
CSV and populates the legacy_id field based on what we were able to
scrape out of the legacy DB. From workshops, we identified the problem
records with

SELECT e.code, p.id, p.email FROM memberships m, people p
  FROM memberships m, people p, events e
  WHERE m.event_id=e.id
  AND m.person_id=p.id
  AND e.start_date>'2022-01-01'
  AND p.legacy_id IS NULL;

To get the legacy DB values, we wrote a script to match the emails from
the above search with the `email1` field of the `People` table.

Signed-off-by: Ian Allison <iana@pims.math.ca>